relocated code from LessonTest to Lesson, removed unused components, renamed SVG to PNG for accuracy, added modal button in translation without adjusting the content yet
