+++ src/components/Navbar.jsx
@@ -1,14 +1,7 @@
 import React, { useEffect } from "react";
 import "./navbar.css";
 import { Link } from "react-router-dom";
-import {
-  NavLogo,
-  HomeSVG,
-  LessonSVG,
-  CharacterSVG,
-  ProfileSVG,
-  AboutSVG,
-} from "./SVG";
+import { NavLogo, LessonPNG, CharacterPNG, ProfilePNG, AboutPNG } from "./PNG";
 import { useAuth0 } from "@auth0/auth0-react";
 import axios from "axios";
 import { Backend_URL } from "../BACKEND_URL.js";
@@ -41,24 +34,18 @@ export default function Navbar(userData) {
       {isAuthenticated ? (
         <ul className="nav-link-container">
           <li className="nav-link-wrapper">
-            <Link to="/" className="nav-link-text-wrapper">
-              <HomeSVG />
-              <p className="nav-link-text">HOME</p>
-            </Link>
-          </li>
-          <li className="nav-link-wrapper">
             <Link
-              to="/vocabs/lesson"
+              to="/lesson"
               className="nav-link-text-wrapper"
               state={{ type: `vocabs` }}
             >
-              <LessonSVG />
+              <LessonPNG />
               <span className="nav-link-text">LESSON</span>
             </Link>
           </li>
           <li className="nav-link-wrapper">
             <Link to="/characters/display" className="nav-link-text-wrapper">
-              <CharacterSVG />
+              <CharacterPNG />
               <span className="nav-link-text">CHARACTERS</span>
             </Link>
           </li>
@@ -71,7 +58,7 @@ export default function Navbar(userData) {
                   className="nav-link-profile-image"
                 />
               ) : (
-                <ProfileSVG />
+                <ProfilePNG />
               )}
 
               <span className="nav-link-text">PROFILE</span>
@@ -79,13 +66,13 @@ export default function Navbar(userData) {
           </li>
           <li className="nav-link-wrapper">
             <Link to="/about/about" className="nav-link-text-wrapper">
-              <AboutSVG />
+              <AboutPNG />
               <span className="nav-link-text">ABOUT</span>
             </Link>
           </li>
           <li className="nav-link-wrapper">
             <Link to="/about/landing" className="nav-link-text-wrapper">
-              <AboutSVG />
+              <AboutPNG />
               <span className="nav-link-text">LANDING</span>
             </Link>
           </li>
