+++ src/components/QuestionsComponents/Matching.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { MiniCharacter } from "../SVG";
+import { MiniCharacter } from "../PNG";
 import "./matching.css";
 import { Button } from "antd";
 import axios from "axios";
@@ -14,52 +14,58 @@ export default function Matching(props) {
   const [rightSelect, setRightSelect] = useState("");
   const [inputRow, setInputRow] = useState([]); // This is the left row
   const [outputRow, setOutputRow] = useState([]); // This is the right row
-  const [count, setCount] = useState(0)
+  const [count, setCount] = useState(0);
   //query random input
   useEffect(() => {
-    axios.post(`${Backend_URL}/questions/matching/random`, {questionData: questionData}).then((res) => {
-      setInputRow(res.data.inputRow)
-      setOutputRow(res.data.outputRow)
-      for(let i = 0; i < 5; i++){
-        disabledTrackerLeft[`${res.data.inputRow[i]}`] = false
-        disabledTrackerRight[`${res.data.outputRow[i]}`] = false;
-      }
-    })
-    setCount(0)
-    disabledTrackerLeft={}
-    disabledTrackerRight={}
+    axios
+      .post(`${Backend_URL}/questions/matching/random`, {
+        questionData: questionData,
+      })
+      .then((res) => {
+        setInputRow(res.data.inputRow);
+        setOutputRow(res.data.outputRow);
+        for (let i = 0; i < 5; i++) {
+          disabledTrackerLeft[`${res.data.inputRow[i]}`] = false;
+          disabledTrackerRight[`${res.data.outputRow[i]}`] = false;
+        }
+      });
+    setCount(0);
+    disabledTrackerLeft = {};
+    disabledTrackerRight = {};
   }, [questionData]);
   //verify
   useEffect(() => {
-    if(leftSelect !== "" && rightSelect !== ""){
-      axios.post(`${Backend_URL}/questions/matching/verify`, {
-        left: leftSelect,
-        right: rightSelect,
-        type: questionData.question_type.split("-")
-      }).then((res) => {
-        if(res.data.isCorrect){
-          disabledTrackerLeft[`${leftSelect}`] = true
-          disabledTrackerRight[`${rightSelect}`] = true
-          setCount(count + 1)
-        }else{
-          //for future changing colour to red
-          console.log("you are wrong")
-        }
-        setLeftSelect("");
-        setRightSelect("");
-      })
+    if (leftSelect !== "" && rightSelect !== "") {
+      axios
+        .post(`${Backend_URL}/questions/matching/verify`, {
+          left: leftSelect,
+          right: rightSelect,
+          type: questionData.question_type.split("-"),
+        })
+        .then((res) => {
+          if (res.data.isCorrect) {
+            disabledTrackerLeft[`${leftSelect}`] = true;
+            disabledTrackerRight[`${rightSelect}`] = true;
+            setCount(count + 1);
+          } else {
+            //for future changing colour to red
+            console.log("you are wrong");
+          }
+          setLeftSelect("");
+          setRightSelect("");
+        });
     }
-  }, [leftSelect, rightSelect])
-  const addLeft = (e) =>{
+  }, [leftSelect, rightSelect]);
+  const addLeft = (e) => {
     const text = e.target.textContent;
-    setLeftSelect(text)
-  }
+    setLeftSelect(text);
+  };
 
   const addRight = (e) => {
     const text = e.target.textContent;
-    setRightSelect(text)
+    setRightSelect(text);
   };
-  console.log(disabledTrackerLeft)
+  console.log(disabledTrackerLeft);
   console.log(disabledTrackerRight);
   //Set output rows
   const leftCol = inputRow?.map((x) => (
@@ -83,8 +89,8 @@ export default function Matching(props) {
     </Button>
   ));
 
-  if(count === 5){
-    props.setHasSubmit(true)
+  if (count === 5) {
+    props.setHasSubmit(true);
   }
 
   return (
@@ -100,4 +106,3 @@ export default function Matching(props) {
     </>
   );
 }
-
