+++ src/components/Start.jsx
@@ -1,15 +1,14 @@
 import React from "react";
-import { MiniCharacter } from "./SVG"
+import { MiniCharacter } from "./PNG";
 
 const Start = (props) => {
-  const placeholder = props.type //Depends on type
+  const placeholder = props.type; //Depends on type
   return (
     <>
       <MiniCharacter />
       <p>Lets start learning some Japanese {placeholder}!</p>
     </>
   );
-}
+};
 
-
-export default Start
+export default Start;
