--- src/components/lessonTest.css
@@ -1,20 +1,0 @@
-.lesson-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  width: 100%;
-}
-
-.lesson-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 60%;
-}
-
-.question-wrapper {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 50px;
-}
