+++ src/pages/EditProfile.jsx
@@ -5,7 +5,7 @@ import { storage } from "../firebase";
 import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import axios from "axios";
 import { Link, Outlet, useOutletContext } from "react-router-dom";
-import { EditBtn } from "../components/SVG";
+import { EditBtn } from "../components/PNG";
 import { Button } from "../components/Buttons";
 
 const PROFILE_PHOTO_FOLDER = "profile-picture-url";
@@ -108,88 +108,86 @@ export default function EditProfile() {
 
   return (
     <>
-      <div className="edit-profile-section">
-        <div className="edit-profile-container">
-          <div className="edit-profile-image-wrapper">
-            <h1 className="edit-profile-title">Edit Profile</h1>
-            <div className="profile-photo-submit">
-              <label className="custom-file-upload">
-                <div className="custom-file-edit">
-                  <EditBtn />
-                </div>
-                {isChangedPhoto === false ? (
-                  <img
-                    src={userData?.profile_pic_url}
-                    alt={userData?.profile_pic_url}
-                    className="edit-profile-image"
-                  />
-                ) : (
-                  <img
-                    src={updatedPhotoFileUrl}
-                    alt={updatedPhotoFileUrl}
-                    className="edit-profile-image"
-                  />
-                )}
-                <input
-                  className="file-input"
-                  type="file"
-                  onChange={handleUpdatedPhoto}
+      <div className="edit-profile-container">
+        <div className="edit-profile-image-wrapper">
+          <h1 className="edit-profile-title">Edit Profile</h1>
+          <div className="profile-photo-submit">
+            <label className="custom-file-upload">
+              <div className="custom-file-edit">
+                <EditBtn />
+              </div>
+              {isChangedPhoto === false ? (
+                <img
+                  src={userData?.profile_pic_url}
+                  alt={userData?.profile_pic_url}
+                  className="edit-profile-image"
                 />
-              </label>
-              <Button onClick={handlePhotoSubmit}>Submit</Button>
-            </div>
+              ) : (
+                <img
+                  src={updatedPhotoFileUrl}
+                  alt={updatedPhotoFileUrl}
+                  className="edit-profile-image"
+                />
+              )}
+              <input
+                className="file-input"
+                type="file"
+                onChange={handleUpdatedPhoto}
+              />
+            </label>
+            <Button onClick={handlePhotoSubmit}>Submit</Button>
           </div>
-          <div className="edit-profile-info-container">
-            <div className="edit-profile-info-wrapper">
-              <h1 className="edit-profile-title-info">First Name:</h1>
-              <input
-                className="edit-profile-text-info"
-                name="first_name"
-                value={textInput.first_name}
-                placeholder="Enter name"
-                onChange={handleTextInputChange}
-              />
-            </div>
-            <div className="edit-profile-info-wrapper">
-              <h1 className="edit-profile-title-info">Last Name:</h1>
-              <input
-                className="edit-profile-text-info"
-                name="last_name"
-                value={textInput.last_name}
-                placeholder="Enter last name"
-                onChange={handleTextInputChange}
-              />
-            </div>
-            <div className="edit-profile-info-wrapper">
-              <h1 className="edit-profile-title-info">Username:</h1>
-              <input
-                className="edit-profile-text-info"
-                name="username"
-                value={textInput.username}
-                placeholder="Enter username"
-                onChange={handleTextInputChange}
-              />
-            </div>
-            <div className="edit-profile-info-wrapper">
-              <h1 className="edit-profile-title-info">Waifu Voice:</h1>
-              <select
-                className="profile-input-box"
-                name="voicevox_id"
-                value={textInput.voicevox_id}
-                onChange={handleTextInputChange}
-              >
-                {voices &&
-                  voices.map((voice, index) => (
-                    <option value={voice.id} key={index}>
-                      {voice.id}
-                    </option>
-                  ))}
-              </select>
-            </div>
-            <Link to="/profile/user">
-              <Button onClick={handleProfileChange}>Done</Button>
-            </Link>
+        </div>
+        <div className="edit-profile-info-container">
+          <div className="edit-profile-info-wrapper">
+            <h1 className="edit-profile-title-info">First Name:</h1>
+            <input
+              className="edit-profile-text-info"
+              name="first_name"
+              value={textInput.first_name}
+              placeholder="Enter name"
+              onChange={handleTextInputChange}
+            />
           </div>
+          <div className="edit-profile-info-wrapper">
+            <h1 className="edit-profile-title-info">Last Name:</h1>
+            <input
+              className="edit-profile-text-info"
+              name="last_name"
+              value={textInput.last_name}
+              placeholder="Enter last name"
+              onChange={handleTextInputChange}
+            />
+          </div>
+          <div className="edit-profile-info-wrapper">
+            <h1 className="edit-profile-title-info">Username:</h1>
+            <input
+              className="edit-profile-text-info"
+              name="username"
+              value={textInput.username}
+              placeholder="Enter username"
+              onChange={handleTextInputChange}
+            />
+          </div>
+          <div className="edit-profile-info-wrapper">
+            <h1 className="edit-profile-title-info">Waifu Voice:</h1>
+            <select
+              className="profile-input-box"
+              name="voicevox_id"
+              value={textInput.voicevox_id}
+              onChange={handleTextInputChange}
+            >
+              {voices &&
+                voices.map((voice, index) => (
+                  <option value={voice.id} key={index}>
+                    {voice.id}
+                  </option>
+                ))}
+            </select>
+          </div>
+          <Link to="/profile/user">
+            <Button onClick={handleProfileChange}>Done</Button>
+          </Link>
         </div>
       </div>
       <div>
