+++ src/pages/characters.css
@@ -1,5 +1,3 @@
-@import "@fontsource/zen-maru-gothic";
-
 .character-container {
   display: flex;
   flex-direction: column;
@@ -12,7 +10,7 @@
 
 .character-selection-title {
   display: flex;
-  width: 50%;
+  width: 70%;
   margin-bottom: 30px;
 }
